Rename "meta" => "config"
