+++ src/hssm/config.py
@@ -13,7 +13,7 @@ from ._types import LogLik, LoglikKind, SupportedModels
 from .defaults import (
     default_model_config,
 )
-from .modelmeta import get_default_model_config
+from .modelconfig import get_default_model_config
 from .register import register_model
 
 if TYPE_CHECKING:
