+++ src/hssm/modelconfig/weibull_config.py
